requirements.txt
openai
Flask
python-dotenv


